Updated AI and added chest depositing.
Added Auto Chest Find & Deposit.
Added Chest blacklisting on timeout condition.
Added storeCarrots with ShiftClickSlotsForFarm in ContainerScreen54Mixin.java which moves all the inventory except a couple stacks of carrots to the target container.
Reorganised AI behaviour.

Final behavior of push. Harvests all plants around the area, replants and picks up carrots (carrots only atm) and when the inventory is above 32 slots, it attempts to deposit all carrots to the closest non-blacklisted chest, if player inventory is not relatively empty by 800 onUpdate passes it blacklists that BlockPos of the chest and moves to the next one.